+++ src/Pages/About/Our Team/index.jsx
@@ -6,7 +6,7 @@ import { FaChevronRight } from "react-icons/fa";
 import Management from "./Management";
 import Board from "./Board";
 import Staff from "./Staff";
-import Navbar from "components/Navbar";
+import MobileSideBar from "components/MobileSideBar";
 
 const OurTeam = () => {
   const menuLinks = [
@@ -37,17 +37,7 @@ const OurTeam = () => {
 
   return (
     <Box>
-      <Navbar
-        display={{
-          sm: "flex",
-          md: "flex",
-          lg: "flex",
-          xl: "none",
-          "2xl": "none",
-        }}
-
-        bg="red"
-      />
+      <MobileSideBar />
 
       <Box
         h="441px"
