+++ src/Pages/About/Profile/Index.jsx
@@ -7,7 +7,7 @@ import Gallery from "Pages/About/Profile/Gallery";
 import OurPhilosophy from "./OurPhilosophy";
 import OurClout from "./OurClout";
 import { useNavigate } from "react-router-dom";
-import Navbar from "components/Navbar";
+import MobileSideBar from "components/MobileSideBar";
 
 const Profile = () => {
   const fileUrl = "/documents/Ojumo House.pdf";
@@ -25,7 +25,8 @@ const Profile = () => {
 
   return (
     <Fragment>
-      <Navbar />
+      <MobileSideBar />
+      
       <Box
         h="441px"
         w="100%"
