+++ src/Pages/Academics/Index.jsx
@@ -9,6 +9,7 @@ import NYSC from "Pages/Academics/NYSC";
 import Footer from "components/Footer";
 import FooterCopywright from "components/FooterCopywright";
 import LinkTabs from "components/LinkTabs";
+import Navbar from "components/Navbar";
 
 const Admission = () => {
   const { id } = useParams();
@@ -58,6 +59,7 @@ const Admission = () => {
 
   return (
     <Fragment>
+      <Navbar />
       <Box
         h="441px"
         w="100%"
