+++ src/Pages/Academics/Programmes.jsx
@@ -14,7 +14,7 @@ import studyProgrammes1 from "assets/Images/studyprogrammes1.png";
 import studyProgrammes3 from "assets/Images/studyprogrammes3.png";
 import studyProgrammes2 from "assets/Images/studyprogrammes2.png";
 import fileDownload from "assets/Images/fileDownload.png";
-import Navbar from "components/Navbar";
+import MobileSideBar from "components/MobileSideBar";
 
 const Programmes = () => {
   const programmes = [
@@ -106,7 +106,7 @@ const Programmes = () => {
 
   return (
     <Box>
-      <Navbar />
+      <MobileSideBar />
       <Box
         h="441px"
         w="100%"
