+++ src/Pages/Admission/FAQs.jsx
@@ -17,7 +17,7 @@ import { FaChevronCircleUp } from "react-icons/fa";
 import React, { Fragment } from "react";
 import waveimg from "assets/Images/waveimg.png";
 import { useNavigate } from "react-router-dom";
-import Navbar from "components/Navbar";
+import MobileSideBar from "components/MobileSideBar";
 
 const FAQs = () => {
   const FAQs = [
@@ -159,7 +159,7 @@ const FAQs = () => {
   ];
   return (
     <Fragment>
-      <Navbar />
+      <MobileSideBar />
       <Box
         h="441px"
         w="100%"
